refactor(core): split files-plugin into focused modules
Break the 1.2k-line files-plugin.ts into single-purpose files and fold
its four HTTP endpoints + watchers into the unified api-plugin so all
dev-server APIs live in one place:

- slide-ops.ts: source-editing helpers (meta.title, page/notes reorder,
  remove, duplicate) + slide path resolvers
- assets.ts: mime lookup, asset-name validation, scoped path resolution
- folders.ts: manifest read/write + folder name/icon validators
- api-plugin.ts: gains /__slides, /__assets, /__svgl, /__folders
  middlewares and the asset/folders watcher
- Tests split into slide-ops, assets, and folders
