+++ packages/core/src/vite/config.ts
@@ -7,7 +7,6 @@ import type { InlineConfig } from 'vite';
 import { apiPlugin } from './api-plugin.ts';
 import { currentPlugin } from './current-plugin.ts';
 import { designPlugin } from './design-plugin.ts';
-import { filesPlugin } from './files-plugin.ts';
 import { locTagsPlugin } from './loc-tags-plugin.ts';
 import { notesPlugin } from './notes-plugin.ts';
 import { loadUserConfig, type OpenSlideConfig, openSlidePlugin } from './open-slide-plugin.ts';
@@ -52,9 +51,8 @@ export async function createViteConfig(opts: CreateViteConfigOptions): Promise<I
       openSlidePlugin({ userCwd, config }),
       themesPlugin({ userCwd, config }),
       designPlugin({ userCwd }),
-      apiPlugin({ userCwd, slidesDir }),
+      apiPlugin({ userCwd, slidesDir, assetsDir }),
       notesPlugin({ userCwd, slidesDir }),
-      filesPlugin({ userCwd, slidesDir, assetsDir }),
       currentPlugin({ userCwd, slidesDir }),
     ],
     resolve: {
